[핀] 체스 스프링 실습 3단계 제출합니다. (#118)
* [알트] 체스 스프링 실습 1단계 제출합니다. (#25)

* docs: README.md 추가

* refactor: 레벨 1 체스 프로그램 개선

* URI 변경
* static file 분리
* JdbcTemplate 리팩토링

* style: 줄바꿈 제거

* refactor: feedback 반영


* [알트] 체스 스프링 실습 2단계 제출합니다. (#65)

* [핀] 체스 스프링 실습 1단계 제출합니다. (#26)

* docs: README.md 추가

* refactor: 레벨 1 체스 프로그램 개선

* URI 변경
* static file 분리
* JdbcTemplate 리팩토링

* style: 줄바꿈 제거

* refactor: feedback 반영

* fix: 오탈자 수정


* feat: spring boot 어플리케이션으로 마이그레이션

* spring을 위한 컨트롤러 추가
* spring bean을 활용하여 객체의 생명주기 관리
* jdbc 연결을 설정하는 방식 개선

* fix: 방이 존재하지 않을 때 메인 화면으로 이동하도록 수정

* fix: 블랙 팀이 화이트로 render되는 이슈 수정

* refactor: semicolon 추가 및 사용되지 않는 인자 제거

* refactor: 불필요한 출력문 제거

* style: override method 최하단으로 이동

* refactor: datasource properties 이름 변경

* refactor: css, js 파일 분리

* refactor: WebConfiguration을 통해 bean을 주입받도록 변경

* Qualifier로 설정하던 기존 방식은 설정 변경이 용이하지 못하다고 판단함
* 한 곳에서 의존 관계를 변경할 수 있도록 WebConfiguration 클래스 추가
* 스프링 내장 DataSource를 읽는 이슈는 WebConfiguration에 annotation을 추가하여 해결
* 두 개 이상의 구현체를 가지는 경우 테스트 클래스에서 특정 구현체를 받도록 수정

* refactor: 불필요한 bean 생성 메서드 제거

* refactor: JdbcConfiguration을 불변 클래스로 변경

* WebConfiguration 이 builder를 통해 DataSource를 생성하도록 변경
* Builder를 bean에서 생성하기 위해 setter 메소드 추가
* WebConfiguration에 존재하는 setter 메소드 제거
* Url과 Option을 합쳐서 반환하는 메서드명 getFullUrl로 변경

* refactor: 불필요한 생성자 제거 및 builder setter 메서드 통합

* builder의 set 네이밍을 가진 메서드를 통합하여 불필요한 코드 중복 제거

* refactor: request body를 dto로 포장

* refactor: template 문자열 사용하도록 변경


* refactor: spring data jdbc로 마이그레이션

* entity 클래스 추가
* id를 Long 타입으로 변경
* spark application 관련 코드 제거
* custom jdbc 관련 코드 제거
* dao 제거
* WebConfiguration.java 제거

* refactor: 불필요한 구문 제거

+++ src/main/java/wooteco/chess/controller/IndexController.java
@@ -21,7 +21,7 @@ public class IndexController {
 	}
 
 	@GetMapping("/rooms/{id}")
-	public String renderRoom(@PathVariable Integer id, Model model) {
+	public String renderRoom(@PathVariable Long id, Model model) {
 		try {
 			chessGameService.find(id);
 			model.addAttribute("id", id);
